🐛 Address PR review issues
- H1: Define local types (Genre, MovieStatus, ProductionCompany, ProductionCountry, SpokenLanguage, MovieCollection) in IntelligenceDomain — removes MoviesDomain dependency violation
- H2: Add Equatable and Sendable to LLMSessionError; change associated Error? values to String? for equatability
- H3: Add fallback catch clause in handleSendPrompt to prevent stuck isThinking state
- M1: Implement errorDescription: String? instead of localizedDescription so LocalizedError machinery works correctly
- M2: Use ISO8601 date format for releaseDate in MovieToolMapper instead of locale-sensitive formatted()
- M3: Expose all new Movie properties in Movie.mock() factory
- M4: Add LLMSessionErrorTests covering all 11 cases plus Equatable and localizedDescription routing
- M5: Replace placeholder MovieIntelligenceFeatureTests with TestStore tests for key state transitions
- L1: Remove commented-out observability code from FoundationModelsLLMSession and FoundationModelsMovieLLMSessionRepository
- L2: Remove self-referential doc comments from Movie.swift
- L3: Change budget/revenue from Double? to Int? in IntelligenceDomain.Movie and MovieToolMovie

+++ Contexts/PopcornIntelligence/Sources/IntelligenceDomain/Entities/Movie.swift
@@ -6,114 +6,36 @@
 //
 
 import Foundation
-import MoviesDomain
 
-///
-/// Represents a movie in the domain model.
-///
-/// A movie contains essential information including its title, overview, release date,
-/// and associated imagery paths. This is the core domain entity used throughout the
-/// movies context for representing movie data.
-///
-/// Represents ``Movie``.
 public struct Movie: Identifiable, Equatable, Sendable {
 
-    /// The unique identifier for the movie.
-    /// The ``id`` value.
     public let id: Int
-
-    /// The movie's title.
-    /// The ``title`` value.
     public let title: String
-
-    /// The ``tagline`` value.
     public let tagline: String?
-
-    /// The ``originalTitle`` value.
     public let originalTitle: String?
-
-    /// The ``originalLanguage`` value.
     public let originalLanguage: String?
-
-    /// A brief description or plot summary of the movie.
-    /// The ``overview`` value.
     public let overview: String
-
-    /// The ``runtime`` value.
     public let runtime: Int?
-
-    /// The ``genres`` value.
     public let genres: [Genre]?
-
-    /// The movie's theatrical release date, if known.
-    /// The ``releaseDate`` value.
     public let releaseDate: Date?
-
-    /// URL path to the movie's poster image.
-    /// The ``posterPath`` value.
     public let posterPath: URL?
-
-    /// URL path to the movie's backdrop image.
-    /// The ``backdropPath`` value.
     public let backdropPath: URL?
-
-    /// The ``budget`` value.
-    public let budget: Double?
-
-    /// The ``revenue`` value.
-    public let revenue: Double?
-
-    /// The ``homepageURL`` value.
+    public let budget: Int?
+    public let revenue: Int?
     public let homepageURL: URL?
-
-    /// The ``imdbID`` value.
     public let imdbID: String?
-
-    /// The ``status`` value.
     public let status: MovieStatus?
-
-    /// The ``productionCompanies`` value.
     public let productionCompanies: [ProductionCompany]?
-
-    /// The ``productionCountries`` value.
     public let productionCountries: [ProductionCountry]?
-
-    /// The ``spokenLanguages`` value.
     public let spokenLanguages: [SpokenLanguage]?
-
-    /// The ``originCountry`` value.
     public let originCountry: [String]?
-
-    /// The ``belongsToCollection`` value.
     public let belongsToCollection: MovieCollection?
-
-    /// The ``popularity`` value.
     public let popularity: Double?
-
-    /// The ``voteAverage`` value.
     public let voteAverage: Double?
-
-    /// The ``voteCount`` value.
     public let voteCount: Int?
-
-    /// The ``hasVideo`` value.
     public let hasVideo: Bool?
-
-    /// The ``isAdultOnly`` value.
     public let isAdultOnly: Bool?
 
-    ///
-    /// Creates a new movie instance.
-    ///
-    /// - Parameters:
-    ///   - id: The unique identifier for the movie.
-    ///   - title: The movie's title.
-    ///   - overview: A brief description or plot summary.
-    ///   - releaseDate: The movie's theatrical release date. Defaults to `nil`.
-    ///   - posterPath: URL path to the poster image. Defaults to `nil`.
-    ///   - backdropPath: URL path to the backdrop image. Defaults to `nil`.
-    ///
-    /// Creates a new instance.
     public init(
         id: Int,
         title: String,
@@ -126,8 +48,8 @@ public struct Movie: Identifiable, Equatable, Sendable {
         releaseDate: Date? = nil,
         posterPath: URL? = nil,
         backdropPath: URL? = nil,
-        budget: Double? = nil,
-        revenue: Double? = nil,
+        budget: Int? = nil,
+        revenue: Int? = nil,
         homepageURL: URL? = nil,
         imdbID: String? = nil,
         status: MovieStatus? = nil,
